Completed the following -------------------------
Angular
--------
UsersService: Accept the typed parameter IUserModel and return a typed return type IUserResultModel when a new user is registered.
ProductsService: Accept a typed parameter IUserModel and return a typed return type for all the methods.
ViewproductComponent: Use the state property of the NavigationExtras class to communicate product details instead of raw JSON data passed as a query string. Additionally, I am only receiving the product name instead of the entire product data model and then making a post-call to the API to retrieve the product details.
ListproductsComponent: Pass the product name to ViewproductComponent when a product is selected using NavigationExtras.
Created a models folder and categorized some of the models into their respective folders.
IProductCategoryModel: Created a new interface which represents a response structure for a unique product category.
IUserResultModel: Created a new interface which represents the result after creation of a new user.   It intentionally does not contain password property for security reasons.
IUserModel: Created a new interface which represents the structure for a unique user.
LoginComponent: While creating a new user registration, a new object of the type IUserModel is created. The response of type IUserResultModel is received instead of a string.
CategoriesComponent: Use the typeed response IProductCategoryModel instead of any type.
ViewcartComponent: Clicking on products from this component will open the ViewproductComponent with the product details.
TokeninterceptorService: Created a new method called handleError and included the error handling logic here.

API
----
1. UserDataProvider: Made changes to the addNewUser method to return the newly created UserID with an output parameter while creating a new user.
2. Folder restructure in the ShoppingCart.API.Models project
3. UserResultModel: Data model that holds the result after a new user is created.This model intentionally does not contain password property for security reasons.
4. UsersV1Controller: Made changes to the endpoint 'PostUser' to return UserResultModel.

Database
----------
Made changes to the following stored procedure
Sch_UserManagement.sp_CreateUser: Returns the newly created user id.